Enable -Werror (#6171)
* Enable -Werror

All warnings are now errors. External libraries are included
appropriately as system includes so that diagnostics are not emitted.
For the pyx libraries, additional flags such as "-Wno-unused-variable"
are supplied.

* [run CI]

* [run CI]

* Silence warning from #including pyfs.cpp

* [run CI]

* Silence warning from #including arrow_ext.cpp

* [run CI]

* [run CI]

* Fix redefined macro

* Fix int->uint64_t

* [run CI]

* Revert "Fix redefined macro"

This reverts commit 1c98a69c79d7dd8fb867ad439d83852bf2711cf6.

* Fix return type

* Allow macro redef

* [run CI]

* undefine macros before redefining them

The warning cannot be disabled on some compilers.

* [run CI]

* Fix a few more int types

* [run CI]

* c-style cast -> static_cast

* [run CI]

+++ bodo/libs/_stream_join.cpp
@@ -637,7 +637,8 @@ void generate_build_table_outer_rows_for_partition(
     // Add unmatched rows from build table to output table
     for (size_t i_row = 0;
          i_row < partition->build_table_buffer.data_table->nrows(); i_row++) {
-        if ((!requires_reduction || ((i_row % n_pes) == my_rank))) {
+        if ((!requires_reduction ||
+             ((i_row % n_pes) == static_cast<size_t>(my_rank)))) {
             bool has_match = GetBit(build_table_matched_->data(), i_row);
             if (!has_match) {
                 build_idxs.push_back(i_row);
@@ -1520,8 +1521,8 @@ std::shared_ptr<table_info> filter_na_values(
         } else if (build_table_outer) {
             // If build table is replicated but output is not
             // replicated evenly divide NA values across all ranks.
-            append_nas[i] =
-                add_all || ((join_state->build_na_counter % n_pes) == myrank);
+            append_nas[i] = add_all || ((join_state->build_na_counter %
+                                         n_pes) == static_cast<size_t>(myrank));
             join_state->build_na_counter++;
         }
     }
@@ -2478,7 +2479,7 @@ uint32_t get_partition_top_bitmask_by_idx(JoinState* join_state, int64_t idx) {
     try {
         std::vector<std::shared_ptr<JoinPartition>>& partitions =
             ((HashJoinState*)join_state)->partitions;
-        if (idx >= partitions.size()) {
+        if (static_cast<size_t>(idx) >= partitions.size()) {
             throw std::runtime_error(
                 "get_partition_top_bitmask_by_idx: partition index " +
                 std::to_string(idx) + " out of bound: " + std::to_string(idx) +
